Added checks to avoid writes random content on pointers
The related issue is Issue #10 in lis2dux12-pid repository (see
https://github.com/STMicroelectronics/lis2dux12-pid/issues/10).

Reads could write random data to pointer since no checks on
the return status are made.

Another set of analogous checks was added to the
membank so it jumps to the exit condition (resetting the
membank to its correct state) before returning the error.

Signed-off-by: Gioele Fiorenza <gioele.fiorenza2000@gmail.com>